This is the initial release.
It was built against the .NET Framework Library 4.5.1. You will likely need that version or higher for this to run properly.

When downloaded, extract and run "SiaConsoleLogPackager.exe" and it will copy all your log files and place them in a folder on your desktop and zip them up for you.

This will be useful when asking for assistance, as many users do not know where the log files are stored, or what specific logs are needed. They are plain text files, so you can clearly see any files that you are copying to share.